Add parameter for configuring environment
If not specified, the old module behaviour is retained, i.e. configure
the environment of the current puppet run into puppet.conf.
This might not always be desirable, e.g. when manually running puppet
towards a temporary test environment.
The env setting allows for statically configuring the environment to
use, e.g. production.