Add parameter env for configuring environment option in puppet.conf
If not specified, the environment of the current puppet run will be used
for the environment option in puppet.conf. This might not always be
desirable, e.g. when manually running puppet with temporary test
environment (puppet agent -t --environment foo). The env parameter
allows for statically configuring the environment to use so that when
specifying the environment from the command line, the option in the
puppet.conf does not change.